Recreated the grayscale_jpeg test case
The original greyscale_jpeg image had the color transform in the
frame set to kNone. The new one to YCbCr.

This partially fixes issue https://github.com/libjxl/libjxl/issues/584

This does not fix the fact that it's uncertain if the behavior of
grayscale kNone vardct is well defined.

A different test case could also be added for kNone grayscale later.

THe newly generated image decodes correctly with djxl and with the
decode_oneshot API example, because kYCbCr's grayscale behavior in jxl
is will defined and implemented.

The image was generated the same way as the previous one, using cjxl
on the jpg image without additional flags, which makes it automatically
add the jpeg reconstruction box. So something must have changed in
cjxl's behavior since then, given that it produced the image with kNone
back then, with kYCbCr now.