Add level 10 jxll boxes to some images
This adds level 10 boxes to those images that use level 10 features (so
not compatible with the specified level 5) to make those conformance
testing images valid.

alpha_premultiplied: because it uses 16-bit alpha combined with squeeze
which gives > 16 bit depth

cmyk_layers: CMYK is only enabled for level 10

lossless_pfm: because it uses 32 bits per channel

In each case, the image was a pure codestream, so JXL, ftyp, jxll
and a unbounded (size 0) jxlc box header are added.